+++ Editor/Program.cs
@@ -1,0 +1,24 @@
+using System;
+using System.Windows.Forms;
+
+using Editor.Forms;
+
+namespace Editor
+{
+    static class Program
+    {
+        /// <summary>
+        ///  The main entry point for the application.
+        /// </summary>
+        [STAThread]
+        static void Main()
+        {
+#if !NET472
+            Application.SetHighDpiMode(HighDpiMode.SystemAware);
+#endif
+            Application.EnableVisualStyles();
+            Application.SetCompatibleTextRenderingDefault(false);
+            Application.Run(new MainForm());
+        }
+    }
+}